backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow>=2.15.0
spotipy==2.23.0
python-dotenv==1.0.0
requests==2.31.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
gunicorn==21.2.0